use ParseRepositoryTag in image
